backend/requirements.txt
# requirements.txt

# Core web framework
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0

# Computer Vision and Machine Learning
opencv-python>=4.8.0
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0

# Server and Environment
python-dotenv>=1.0.0
gunicorn>=21.2.0
websockets>=12.0
daphne>=4.0.0

# Additional dependencies
asgiref>=3.8.1
sqlparse>=0.3.1
channels_redis>=4.1.0